refactor(security): split policy.rs (1426 lines) into policy/ directory
Breaks the monolithic policy.rs into five focused modules under
src/openhuman/security/policy/:
- types.rs       — enums, structs, ActionTracker, workspace-internal lists
- command_checks.rs — command risk/class classification and allowlist logic
- path_checks.rs — path validation, symlink safety, workspace/trusted-root checks
- enforcement.rs — gate decisions, tool-op enforcement, rate limiting, from_config
- mod.rs         — re-exports preserving the existing public API surface

policy_command.rs and policy_tests.rs moved into the directory (rename detected).
External API unchanged; cargo check passes clean.
